backend/requirements.txt
# Core
openai>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
chromadb>=0.5.0
python-dotenv>=1.0.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# CLI
typer>=0.12.0
rich>=13.7.0

# Telegram
python-telegram-bot>=21.0
